feat: add covariance resampling to `f_0`, `a_0`, `f_2`, and `a_2` K-matrix amplitudes
