+++ crates/laddu-amplitudes/src/kmatrix/f2.rs
@@ -1,0 +1,360 @@
+use nalgebra::{matrix, vector};
+use nalgebra::{SMatrix, SVector};
+use serde::{Deserialize, Serialize};
+use std::array;
+
+use laddu_core::{
+    amplitudes::{Amplitude, AmplitudeID, ParameterLike},
+    data::Event,
+    resources::{Cache, ComplexVectorID, MatrixID, ParameterID, Parameters, Resources},
+    utils::variables::{Mass, Variable},
+    Complex, DVector, Float, LadduError,
+};
+
+#[cfg(feature = "python")]
+use laddu_python::{
+    amplitudes::{PyAmplitude, PyParameterLike},
+    utils::variables::PyMass,
+};
+#[cfg(feature = "python")]
+use pyo3::prelude::*;
+
+use super::FixedKMatrix;
+
+const G_F2: SMatrix<Float, 4, 4> = matrix![
+     0.40033,  0.01820, -0.06709, -0.49924;
+     0.15479,  0.17300,  0.22941,  0.19295;
+    -0.08900,  0.32393, -0.43133,  0.27975;
+    -0.00113,  0.15256,  0.23721, -0.03987
+];
+
+const C_F2: SMatrix<Float, 4, 4> = matrix![
+    -0.04319,  0.00000,  0.00984,  0.01028;
+     0.00000,  0.00000,  0.00000,  0.00000;
+     0.00984,  0.00000, -0.07344,  0.05533;
+     0.01028,  0.00000,  0.05533, -0.05183
+];
+
+const M_F2: SVector<Float, 4> = vector![1.15299, 1.48359, 1.72923, 1.96700];
+
+const COV_F2: SMatrix<Float, 36, 36> = matrix![
+    0.00000, -0.00000, 0.00000, 0.00000, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, -0.00000, 0.00000;
+    -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, -0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, -0.00000, -0.00000, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000;
+    0.00000, 0.00000, 0.00021, 0.00000, -0.00002, 0.00000, -0.00049, 0.00000, -0.00000, 0.00010, 0.00052, 0.00000, -0.00001, -0.00015, -0.00113, 0.00000, 0.00000, 0.00000, 0.00000, -0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00001, -0.00001, 0.00000, 0.00000, 0.00000, 0.00002, 0.00000, -0.00001, -0.00008, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    -0.00000, -0.00000, -0.00002, 0.00000, 0.00001, 0.00000, 0.00007, 0.00000, 0.00000, -0.00002, -0.00006, 0.00000, 0.00000, 0.00002, 0.00012, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00001, -0.00000, 0.00000, -0.00000, 0.00000;
+    0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, -0.00000, -0.00000, 0.00000;
+    -0.00000, -0.00000, -0.00049, 0.00000, 0.00007, -0.00000, 0.00161, 0.00000, 0.00000, -0.00032, -0.00147, 0.00000, -0.00001, 0.00061, 0.00272, 0.00000, -0.00000, 0.00000, 0.00002, 0.00003, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00006, 0.00004, 0.00000, 0.00000, 0.00000, 0.00004, -0.00001, 0.00005, -0.00003, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    -0.00000, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, -0.00000, 0.00010, 0.00000, -0.00002, 0.00000, -0.00032, 0.00000, -0.00000, 0.00027, 0.00028, 0.00000, -0.00001, -0.00011, -0.00054, 0.00000, 0.00000, 0.00000, -0.00001, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00002, 0.00000, 0.00000, 0.00000, 0.00005, 0.00000, -0.00004, -0.00001, 0.00000;
+    -0.00000, 0.00000, 0.00052, 0.00000, -0.00006, 0.00000, -0.00147, 0.00000, 0.00000, 0.00028, 0.00277, 0.00000, -0.00006, -0.00071, -0.00242, 0.00000, 0.00000, 0.00000, -0.00003, -0.00009, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00006, -0.00014, 0.00000, 0.00000, 0.00000, 0.00007, 0.00001, 0.00006, 0.00012, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    -0.00000, -0.00000, -0.00001, 0.00000, 0.00000, -0.00000, -0.00001, 0.00000, -0.00000, -0.00001, -0.00006, 0.00000, 0.00001, 0.00003, 0.00001, 0.00000, -0.00000, 0.00000, 0.00000, 0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00001, 0.00000, 0.00000, 0.00000, -0.00002, -0.00000, -0.00000, -0.00001, 0.00000;
+    -0.00000, -0.00000, -0.00015, 0.00000, 0.00002, 0.00000, 0.00061, 0.00000, 0.00000, -0.00011, -0.00071, 0.00000, 0.00003, 0.00080, 0.00104, 0.00000, -0.00000, 0.00000, 0.00002, 0.00003, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00013, 0.00005, 0.00000, 0.00000, 0.00000, -0.00001, -0.00000, 0.00002, -0.00016, 0.00000;
+    0.00000, -0.00000, -0.00113, 0.00000, 0.00012, -0.00000, 0.00272, 0.00000, 0.00000, -0.00054, -0.00242, 0.00000, 0.00001, 0.00104, 0.00768, 0.00000, -0.00001, 0.00000, -0.00002, 0.00008, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00006, 0.00004, 0.00000, 0.00000, 0.00000, -0.00000, -0.00002, 0.00009, 0.00017, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, -0.00000, 0.00000, 0.00000, -0.00000, -0.00000, -0.00000, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, -0.00001, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, -0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00002, 0.00000, 0.00000, -0.00001, -0.00003, 0.00000, 0.00000, 0.00002, -0.00002, 0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00001, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, -0.00001, 0.00000;
+    0.00000, -0.00000, -0.00001, 0.00000, 0.00000, -0.00000, 0.00003, 0.00000, -0.00000, -0.00000, -0.00009, 0.00000, 0.00001, 0.00003, 0.00008, 0.00000, -0.00000, 0.00000, 0.00000, 0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00001, 0.00001, 0.00000, 0.00000, 0.00000, -0.00001, -0.00000, -0.00001, -0.00002, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, -0.00000, -0.00001, 0.00000, 0.00000, 0.00000, 0.00006, 0.00000, 0.00000, 0.00000, -0.00006, 0.00000, 0.00000, 0.00013, 0.00006, 0.00000, -0.00000, 0.00000, 0.00001, 0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00014, -0.00002, 0.00000, 0.00000, 0.00000, 0.00001, -0.00000, -0.00001, 0.00000, 0.00000;
+    0.00000, -0.00000, -0.00001, 0.00000, 0.00000, -0.00000, 0.00004, 0.00000, 0.00000, -0.00002, -0.00014, 0.00000, 0.00001, 0.00005, 0.00004, 0.00000, -0.00000, 0.00000, 0.00000, 0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00002, 0.00006, 0.00000, 0.00000, 0.00000, -0.00008, -0.00000, -0.00001, -0.00002, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+    -0.00000, 0.00000, 0.00002, 0.00000, -0.00001, -0.00000, 0.00004, 0.00000, -0.00000, 0.00005, 0.00007, 0.00000, -0.00002, -0.00001, -0.00000, 0.00000, -0.00000, 0.00000, -0.00000, -0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00001, -0.00008, 0.00000, 0.00000, 0.00000, 0.00029, 0.00000, -0.00000, -0.00009, 0.00000;
+    -0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, -0.00001, 0.00000, -0.00000, 0.00000, 0.00001, 0.00000, -0.00000, -0.00000, -0.00002, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00000, 0.00000, 0.00000;
+    0.00000, 0.00000, -0.00001, 0.00000, 0.00000, -0.00000, 0.00005, 0.00000, 0.00000, -0.00004, 0.00006, 0.00000, -0.00000, 0.00002, 0.00009, 0.00000, -0.00000, 0.00000, 0.00000, -0.00001, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00001, -0.00001, 0.00000, 0.00000, 0.00000, -0.00000, -0.00000, 0.00003, 0.00002, 0.00000;
+    -0.00000, -0.00000, -0.00008, 0.00000, -0.00000, -0.00000, -0.00003, 0.00000, 0.00000, -0.00001, 0.00012, 0.00000, -0.00001, -0.00016, 0.00017, 0.00000, 0.00000, 0.00000, -0.00001, -0.00002, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, -0.00002, 0.00000, 0.00000, 0.00000, -0.00009, 0.00000, 0.00002, 0.00046, 0.00000;
+    0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000, 0.00000;
+];
+
+/// A K-matrix parameterization for $`f_2`$ particles described by Kopf et al.[^1] with fixed couplings and mass poles
+/// (free production couplings only).
+///
+/// [^1]: Kopf, B., Albrecht, M., Koch, H., Küßner, M., Pychy, J., Qin, X., & Wiedner, U. (2021). Investigation of the lightest hybrid meson candidate with a coupled-channel analysis of $`\bar{p}p`$-, $`\pi^- p`$- and $`\pi \pi`$-Data. The European Physical Journal C, 81(12). [doi:10.1140/epjc/s10052-021-09821-2](https://doi.org/10.1140/epjc/s10052-021-09821-2)
+#[derive(Clone, Serialize, Deserialize)]
+pub struct KopfKMatrixF2 {
+    name: String,
+    channel: usize,
+    mass: Mass,
+    constants: FixedKMatrix<4, 4>,
+    couplings_real: [ParameterLike; 4],
+    couplings_imag: [ParameterLike; 4],
+    couplings_indices_real: [ParameterID; 4],
+    couplings_indices_imag: [ParameterID; 4],
+    ikc_cache_index: ComplexVectorID<4>,
+    p_vec_cache_index: MatrixID<4, 4>,
+}
+
+impl KopfKMatrixF2 {
+    /// Construct a new [`KopfKMatrixF2`] with the given name, production couplings, channel,
+    /// and input mass.
+    ///
+    /// | Channel index | Channel |
+    /// | ------------- | ------- |
+    /// | 0             | $`\pi\pi`$ |
+    /// | 1             | $`2\pi 2\pi`$ |
+    /// | 2             | $`K\bar{K}`$ |
+    /// | 3             | $`\eta\eta`$ |
+    ///
+    /// | Pole names |
+    /// | ---------- |
+    /// | $`f_2(1270)`$ |
+    /// | $`f_2'(1525)`$ |
+    /// | $`f_2(1810)`$ |
+    /// | $`f_2(1950)`$ |
+    pub fn new(
+        name: &str,
+        couplings: [[ParameterLike; 2]; 4],
+        channel: usize,
+        mass: &Mass,
+        seed: Option<usize>,
+    ) -> Box<Self> {
+        let mut couplings_real: [ParameterLike; 4] = array::from_fn(|_| ParameterLike::default());
+        let mut couplings_imag: [ParameterLike; 4] = array::from_fn(|_| ParameterLike::default());
+        for i in 0..4 {
+            couplings_real[i] = couplings[i][0].clone();
+            couplings_imag[i] = couplings[i][1].clone();
+        }
+        Self {
+            name: name.to_string(),
+            channel,
+            mass: mass.clone(),
+            constants: FixedKMatrix::new(
+                G_F2,
+                C_F2,
+                vector![0.1349768, 2.0 * 0.1349768, 0.493677, 0.547862],
+                vector![0.1349768, 2.0 * 0.1349768, 0.497611, 0.547862],
+                M_F2,
+                None,
+                2,
+                COV_F2,
+                seed,
+            ),
+            couplings_real,
+            couplings_imag,
+            couplings_indices_real: [ParameterID::default(); 4],
+            couplings_indices_imag: [ParameterID::default(); 4],
+            ikc_cache_index: ComplexVectorID::default(),
+            p_vec_cache_index: MatrixID::default(),
+        }
+        .into()
+    }
+}
+
+#[typetag::serde]
+impl Amplitude for KopfKMatrixF2 {
+    fn register(&mut self, resources: &mut Resources) -> Result<AmplitudeID, LadduError> {
+        for i in 0..self.couplings_indices_real.len() {
+            self.couplings_indices_real[i] = resources.register_parameter(&self.couplings_real[i]);
+            self.couplings_indices_imag[i] = resources.register_parameter(&self.couplings_imag[i]);
+        }
+        self.ikc_cache_index = resources
+            .register_complex_vector(Some(&format!("KopfKMatrixF2<{}> ikc_vec", self.name)));
+        self.p_vec_cache_index =
+            resources.register_matrix(Some(&format!("KopfKMatrixF2<{}> p_vec", self.name)));
+        resources.register_amplitude(&self.name)
+    }
+
+    fn precompute(&self, event: &Event, cache: &mut Cache) {
+        let s = self.mass.value(event).powi(2);
+        cache.store_complex_vector(
+            self.ikc_cache_index,
+            self.constants.ikc_inv_vec(s, self.channel),
+        );
+        cache.store_matrix(self.p_vec_cache_index, self.constants.p_vec_constants(s));
+    }
+
+    fn compute(&self, parameters: &Parameters, _event: &Event, cache: &Cache) -> Complex<Float> {
+        let betas = SVector::from_fn(|i, _| {
+            Complex::new(
+                parameters.get(self.couplings_indices_real[i]),
+                parameters.get(self.couplings_indices_imag[i]),
+            )
+        });
+        let ikc_inv_vec = cache.get_complex_vector(self.ikc_cache_index);
+        let p_vec_constants = cache.get_matrix(self.p_vec_cache_index);
+        FixedKMatrix::compute(&betas, &ikc_inv_vec, &p_vec_constants)
+    }
+
+    fn compute_gradient(
+        &self,
+        _parameters: &Parameters,
+        _event: &Event,
+        cache: &Cache,
+        gradient: &mut DVector<Complex<Float>>,
+    ) {
+        let ikc_inv_vec = cache.get_complex_vector(self.ikc_cache_index);
+        let p_vec_constants = cache.get_matrix(self.p_vec_cache_index);
+        let internal_gradient = FixedKMatrix::compute_gradient(&ikc_inv_vec, &p_vec_constants);
+        for i in 0..4 {
+            if let ParameterID::Parameter(index) = self.couplings_indices_real[i] {
+                gradient[index] = internal_gradient[i];
+            }
+            if let ParameterID::Parameter(index) = self.couplings_indices_imag[i] {
+                gradient[index] = Complex::<Float>::I * internal_gradient[i];
+            }
+        }
+    }
+}
+
+/// A fixed K-Matrix Amplitude for :math:`f_2` mesons
+///
+/// Parameters
+/// ----------
+/// name : str
+///     The Amplitude name
+/// couplings : list of list of laddu.ParameterLike
+///     Each initial-state coupling (as a list of pairs of real and imaginary parts)
+/// channel : int
+///     The channel onto which the K-Matrix is projected
+/// mass: laddu.Mass
+///     The total mass of the resonance
+/// seed: int, optional
+///     Seed used to resample fixed K-matrix components according to their covariance
+///     No resampling is done if seed is None
+///
+/// Returns
+/// -------
+/// laddu.Amplitude
+///     An Amplitude which can be registered by a laddu.Manager
+///
+/// See Also
+/// --------
+/// laddu.Manager
+///
+/// Notes
+/// -----
+/// This Amplitude follows the prescription of [Kopf]_ and fixes the K-Matrix to data
+/// from that paper, leaving the couplings to the initial state free
+///
+/// +---------------+-------------------+
+/// | Channel index | Channel           |
+/// +===============+===================+
+/// | 0             | :math:`\pi\pi`    |
+/// +---------------+-------------------+
+/// | 1             | :math:`2\pi 2\pi` |
+/// +---------------+-------------------+
+/// | 2             | :math:`K\bar{K}`  |
+/// +---------------+-------------------+
+/// | 3             | :math:`\eta\eta`  |
+/// +---------------+-------------------+
+///
+/// +---------------------+
+/// | Pole names          |
+/// +=====================+
+/// | :math:`f_2(1270)`   |
+/// +---------------------+
+/// | :math:`f_2'(1525)`  |
+/// +---------------------+
+/// | :math:`f_2(1810)`   |
+/// +---------------------+
+/// | :math:`f_2(1950)`   |
+/// +---------------------+
+///
+#[cfg(feature = "python")]
+#[pyfunction(name = "KopfKMatrixF2", signature = (name, couplings, channel, mass, seed = None))]
+pub fn py_kopf_kmatrix_f2(
+    name: &str,
+    couplings: [[PyParameterLike; 2]; 4],
+    channel: usize,
+    mass: PyMass,
+    seed: Option<usize>,
+) -> PyAmplitude {
+    PyAmplitude(KopfKMatrixF2::new(
+        name,
+        array::from_fn(|i| array::from_fn(|j| couplings[i][j].clone().0)),
+        channel,
+        &mass.0,
+        seed,
+    ))
+}
+
+#[cfg(test)]
+mod tests {
+    // Note: These tests are not exhaustive, they only check one channel
+    use std::sync::Arc;
+
+    use super::*;
+    use approx::assert_relative_eq;
+    use laddu_core::{data::test_dataset, parameter, Manager, Mass};
+    #[test]
+    fn test_f2_evaluation() {
+        let mut manager = Manager::default();
+        let res_mass = Mass::new([2, 3]);
+        let amp = KopfKMatrixF2::new(
+            "f2",
+            [
+                [parameter("p0"), parameter("p1")],
+                [parameter("p2"), parameter("p3")],
+                [parameter("p4"), parameter("p5")],
+                [parameter("p6"), parameter("p7")],
+            ],
+            1,
+            &res_mass,
+            None,
+        );
+        let aid = manager.register(amp).unwrap();
+
+        let dataset = Arc::new(test_dataset());
+        let expr = aid.into();
+        let model = manager.model(&expr);
+        let evaluator = model.load(&dataset);
+
+        let result = evaluator.evaluate(&[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8]);
+
+        assert_relative_eq!(result[0].re, 0.02523304, epsilon = Float::EPSILON.sqrt());
+        assert_relative_eq!(result[0].im, 0.39712393, epsilon = Float::EPSILON.sqrt());
+    }
+
+    #[test]
+    fn test_f2_gradient() {
+        let mut manager = Manager::default();
+        let res_mass = Mass::new([2, 3]);
+        let amp = KopfKMatrixF2::new(
+            "f2",
+            [
+                [parameter("p0"), parameter("p1")],
+                [parameter("p2"), parameter("p3")],
+                [parameter("p4"), parameter("p5")],
+                [parameter("p6"), parameter("p7")],
+            ],
+            1,
+            &res_mass,
+            None,
+        );
+        let aid = manager.register(amp).unwrap();
+
+        let dataset = Arc::new(test_dataset());
+        let expr = aid.into();
+        let model = manager.model(&expr);
+        let evaluator = model.load(&dataset);
+
+        let result = evaluator.evaluate_gradient(&[0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8]);
+
+        assert_relative_eq!(result[0][0].re, -0.3078948, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][0].im, 0.3808689, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][1].re, -result[0][0].im);
+        assert_relative_eq!(result[0][1].im, result[0][0].re);
+        assert_relative_eq!(result[0][2].re, 0.4290085, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][2].im, 0.0799660, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][3].re, -result[0][2].im);
+        assert_relative_eq!(result[0][3].im, result[0][2].re);
+        assert_relative_eq!(result[0][4].re, 0.1657487, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][4].im, -0.0041382, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][5].re, -result[0][4].im);
+        assert_relative_eq!(result[0][5].im, result[0][4].re);
+        assert_relative_eq!(result[0][6].re, 0.0594691, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][6].im, 0.1143819, epsilon = Float::EPSILON.cbrt());
+        assert_relative_eq!(result[0][7].re, -result[0][6].im);
+        assert_relative_eq!(result[0][7].im, result[0][6].re);
+    }
+}
