+++ py-laddu/laddu/amplitudes/kmatrix.pyi
@@ -3,6 +3,7 @@ from typing import Literal, overload
 from laddu.amplitudes import Amplitude, ParameterLike
 from laddu.utils.variables import Mass
 
+
 @overload
 def KopfKMatrixF0(
     name: str,
@@ -15,6 +16,8 @@ def KopfKMatrixF0(
     ],
     channel: Literal[0, 1, 2, 3, 4],
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixF0(
@@ -28,6 +31,8 @@ def KopfKMatrixF0(
     ],
     channel: int,
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixF2(
@@ -40,6 +45,8 @@ def KopfKMatrixF2(
     ],
     channel: Literal[0, 1, 2, 3],
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixF2(
@@ -52,6 +59,8 @@ def KopfKMatrixF2(
     ],
     channel: int,
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixA0(
@@ -62,6 +71,8 @@ def KopfKMatrixA0(
     ],
     channel: Literal[0, 1],
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixA0(
@@ -72,6 +83,8 @@ def KopfKMatrixA0(
     ],
     channel: int,
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixA2(
@@ -82,6 +95,8 @@ def KopfKMatrixA2(
     ],
     channel: Literal[0, 1, 2],
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixA2(
@@ -92,6 +107,8 @@ def KopfKMatrixA2(
     ],
     channel: int,
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixRho(
@@ -102,6 +119,8 @@ def KopfKMatrixRho(
     ],
     channel: Literal[0, 1, 2],
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixRho(
@@ -112,6 +131,8 @@ def KopfKMatrixRho(
     ],
     channel: int,
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixPi1(
@@ -119,6 +140,8 @@ def KopfKMatrixPi1(
     couplings: tuple[tuple[ParameterLike, ParameterLike],],
     channel: Literal[0, 1],
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
 @overload
 def KopfKMatrixPi1(
@@ -126,4 +149,6 @@ def KopfKMatrixPi1(
     couplings: tuple[tuple[ParameterLike, ParameterLike],],
     channel: int,
     mass: Mass,
+    *,
+    seed: int | None = None,
 ) -> Amplitude: ...
